feat: user 타입 추가 및 이름 변경
